requirements.txt
fastapi
uvicorn
requests
pymongo
openai
python-dotenv
pillow
gtts
redis